+++ src/components/Layout/Header.tsx
@@ -1,28 +1,40 @@
-import { useRecoilValue, useResetRecoilState } from 'recoil'
-import store from 'store'
-
-import { currentUserState } from 'states'
-import Button from 'components/Button'
+import { Link, useLocation, useParams } from 'react-router-dom'
 
 import styles from './layout.module.scss'
 
-const Header = () => {
-  const currentUser = useRecoilValue(currentUserState)
-  const resetCurrentUser = useResetRecoilState(currentUserState)
+const PageTitle = () => {
+  const { id } = useParams()
+  const location = useLocation()
+  const currentLocation = location.pathname
 
-  const handleLogoutButton = () => {
-    resetCurrentUser()
-    store.remove('loginData')
+  const makeTitle = () => {
+    if (currentLocation === '/') return '대시보드'
+    if (currentLocation === '/manage') return '회원 관리'
+    return '회원 상세 정보'
   }
+
+  const title = makeTitle()
+
   return (
-    <header className={styles.header}>
-      <h1>Back Office</h1>
-      <div>
-        <span>{currentUser.id}님</span>
-        <Button value='로그아웃' size='medium' onClick={handleLogoutButton} />
+    <header>
+      <div className={styles.menuLink}>
+        <Link to='/'>홈</Link>
+        {currentLocation.startsWith('/manage') && (
+          <>
+            <span className={styles.arrow}>&gt;</span>
+            <Link to='/manage'>회원 관리</Link>
+          </>
+        )}
+        {!!id && (
+          <>
+            <span className={styles.arrow}>&gt;</span>
+            <span>회원 상세</span>
+          </>
+        )}
       </div>
+      <h2 className='title'>{title}</h2>
     </header>
   )
 }
 
-export default Header
+export default PageTitle
