--- src/components/PageTitle/pageTitle.module.scss
@@ -1,14 +1,0 @@
-@use '/src/styles/constants/colors';
-
-.menuLink {
-  margin-top: 40px;
-  color: colors.$GRAY800;
-
-  a {
-    color: colors.$GRAY800;
-  }
-
-  .arrow {
-    margin: 0 8px;
-  }
-}
